// projects/codx-ac/src/lib/vouchers/salesinvoices/interfaces/ISalesInvoicesLine.interface.ts
export interface ISalesInvoicesLine {
  recID: string;
  transID: string;
  rowNo: number;
  lineType: string;
  lineStatus: string;
  itemID: string;
  lotID: string;
  idiMID: string;
  idiM0: string;
  idiM1: string;
  idiM2: string;
  idiM3: string;
  idiM4: string;
  idiM5: string;
  idiM6: string;
  idiM7: string;
  idiM8: string;
  idiM9: string;
  fixedDIMs: string;
  cwum: string;
  cwConversion: number;
  cwQty: number;
  cwPrice: boolean;
  umid: string;
  conversion: number;
  quantity: number;
  costPrice: number;
  costAmt: number;
  salesPrice: number;
  salesAmt: number | null;
  lineDiscPct: number;
  lineDiscAmt: number;
  totalDiscPct: number;
  totalDiscAmt: number;
  discPct: number;
  discAmt: number;
  netAmt: number;
  miscPrice: number;
  miscAmt: number;
  salesTaxPct: number;
  salesTaxAmt: number;
  exciseTaxPct: number;
  exciseTaxAmt: number;
  commissionPct: number;
  commissionAmt: number;
  vatid: string;
  vatBase: number;
  vatPct: number | null;
  vatAmt: number;
  totalAmt: number;
  salesAmt2: number | null;
  discAmt2: number;
  netAmt2: number;
  miscAmt2: number;
  salesTaxAmt2: number;
  exciseTaxAmt2: number;
  commissionAmt2: number;
  vatBase2: number;
  vatAmt2: number;
  totalAmt2: number;
  vatBase3: number;
  vatAmt3: number;
  diM1: string;
  diM2: string;
  diM3: string;
  projectID: string;
  reasonID: string;
  pmtTermID: string;
  consumption: boolean;
  assetType: string;
  assetGroupID: string;
  serviceDate: string | null;
  servicePeriods: number | null;
  employeeID: string;
  siteID: string;
  inventAcctID: string;
  costAcctID: string;
  salesAcctID: string;
  aRAcctID: string;
  note: string;
  refType: string;
  refNo: string;
  refID: string | null;
  refLotID: string | null;
  refLineID: string | null;
  pickedQty: number;
  deliveredQty: number;
  cwPickedQty: number;
  cwDeliveredQty: number;
  autoCreated: boolean;
  createdOn: Date;
  createdBy: string;
  modifiedOn: string | null;
  modifiedBy: string;
  itemName: string;

  _rowIndex?: any;
}